Implement tabs
Changes made:
 - Implemented 3 main tabs in UI - Home, Groups and Posts visible after signing in
