+++ app/(app)/(tabs)/(groups)/_layout.tsx
@@ -1,12 +1,12 @@
 import { Slot, Stack } from "expo-router";
 
 // Import your global CSS file
-import "../../global.css"
+import "../../../../global.css";
 
 export default function _layout() {
   return (
     <Stack>
-      <Stack.Screen name="(tabs)" options={{headerShown: false}} />
+      <Stack.Screen name="index" options={{headerShown: false}} />
     </Stack>
   )
 }
