+++ app/(app)/(tabs)/(posts)/_layout.tsx
@@ -1,0 +1,13 @@
+import { Slot, Stack } from "expo-router";
+
+// Import your global CSS file
+import "../../../../global.css";
+
+export default function _layout() {
+  return (
+    <Stack>
+      <Stack.Screen name="index" options={{headerShown: false}} />
+      <Stack.Screen name="createPost" options={{headerShown: false}} /> 
+    </Stack>
+  )
+}
